scripts/requirements.txt
Pillow
pillow-heif
piexif
